# swagger.json
{
	"components": {
...
		"schemas": {
			"ERole": {
				"enum": [
					"admin",
					"user",
					"guest"
				],
				"type": "string"
			},
...
			"IRoleDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/ERole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"phone": {
						"type": "string"
					},
					"challenge": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"role": {
						"$ref": "#/components/schemas/IRoleDto"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ETeamRole": {
				"enum": [
					"leader",
					"member"
				],
				"type": "string"
			},
			"ITeamMemberDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/IUserDto"
					},
					"role": {
						"$ref": "#/components/schemas/ETeamRole"
					}
				},
				"required": [
					"id",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITeamDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/ITeamMemberDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITeamListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ITeamDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITeamCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"userIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITeamUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"userIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EProjectVisibility": {
				"enum": [
					"private",
					"team",
					"public"
				],
				"type": "string"
			},
			"IProjectDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"key": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"visibility": {
						"$ref": "#/components/schemas/EProjectVisibility"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/ITeamDto"
						},
						"type": "array"
					},
					"lead": {
						"$ref": "#/components/schemas/IUserDto"
					}
				},
				"required": [
					"id",
					"name",
					"key",
					"visibility",
					"createdAt",
					"updatedAt",
					"teams"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IStatusDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"isInitial": {
						"type": "boolean"
					},
					"isFinal": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"color",
					"isInitial",
					"isFinal",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStatusListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IStatusDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStatusCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"isInitial": {
						"type": "boolean"
					},
					"isFinal": {
						"type": "boolean"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStatusUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"isInitial": {
						"type": "boolean"
					},
					"isFinal": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EBoardType": {
				"enum": [
					"scrum",
					"kanban"
				],
				"type": "string"
			},
			"ESprintStatus": {
				"enum": [
					"future",
					"active",
					"closed"
				],
				"type": "string"
			},
			"IBoardDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EBoardType"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"project": {
						"$ref": "#/components/schemas/IProjectDto"
					},
					"sprints": {
						"items": {
							"$ref": "#/components/schemas/ISprintDto"
						},
						"type": "array"
					},
					"activeSprint": {
						"$ref": "#/components/schemas/ISprintDto"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EIssueType": {
				"enum": [
					"story",
					"task",
					"bug",
					"epic",
					"subtask"
				],
				"type": "string"
			},
			"IIssueTypeDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/EIssueType"
					},
					"icon": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"icon",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EPriority": {
				"enum": [
					"highest",
					"high",
					"medium",
					"low",
					"lowest"
				],
				"type": "string"
			},
			"IPriorityDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/EPriority"
					},
					"icon": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"icon",
					"color",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISprintDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"goal": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/ESprintStatus"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"board": {
						"$ref": "#/components/schemas/IBoardDto"
					},
					"issues": {
						"items": {
							"$ref": "#/components/schemas/IIssueDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"startDate",
					"endDate",
					"status",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIssueDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"key": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"estimate": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"$ref": "#/components/schemas/IIssueTypeDto"
					},
					"priority": {
						"$ref": "#/components/schemas/IPriorityDto"
					},
					"status": {
						"$ref": "#/components/schemas/IStatusDto"
					},
					"project": {
						"$ref": "#/components/schemas/IProjectDto"
					},
					"board": {
						"$ref": "#/components/schemas/IBoardDto"
					},
					"sprint": {
						"$ref": "#/components/schemas/ISprintDto"
					},
					"assignee": {
						"$ref": "#/components/schemas/IUserDto"
					},
					"reporter": {
						"$ref": "#/components/schemas/IUserDto"
					},
					"parent": {
						"$ref": "#/components/schemas/IIssueDto"
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/ICommentDto"
						},
						"type": "array"
					},
					"labels": {
						"items": {
							"$ref": "#/components/schemas/ILabelDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"key",
					"title",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICommentDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"author": {
						"$ref": "#/components/schemas/IUserDto"
					},
					"issue": {
						"$ref": "#/components/schemas/IIssueDto"
					}
				},
				"required": [
					"id",
					"text",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILabelDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"color",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ISprintListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ISprintDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISprintCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"goal": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"boardId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"startDate",
					"endDate",
					"boardId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISprintUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"goal": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"boardId": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/ESprintStatus"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ESprintPlanningStatus": {
				"enum": [
					"planning",
					"planned",
					"active",
					"closed"
				],
				"type": "string"
			},
			"ISprintPlanningDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/ESprintPlanningStatus"
					},
					"plannedVelocity": {
						"type": "number",
						"format": "double"
					},
					"actualVelocity": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"sprint": {
						"$ref": "#/components/schemas/ISprintDto"
					},
					"participants": {
						"items": {
							"$ref": "#/components/schemas/IUserDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"status",
					"createdAt",
					"updatedAt",
					"participants"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EPlanningItemStatus": {
				"enum": [
					"proposed",
					"approved",
					"rejected",
					"committed"
				],
				"type": "string"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/team/list": {
			"get": {
				"operationId": "GetTeams",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITeamListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/team/{id}": {
			"get": {
				"operationId": "GetTeamById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITeamDto"
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTeam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITeamDto"
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITeamUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTeam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/team": {
			"post": {
				"operationId": "CreateTeam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITeamDto"
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITeamCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/team/{teamId}/members/{userId}": {
			"post": {
				"operationId": "AddMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITeamDto"
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "teamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"default": "member",
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITeamDto"
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "teamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/project/{projectId}/teams/{teamId}": {
			"post": {
				"operationId": "AddTeamToProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProjectDto"
								}
							}
						}
					}
				},
				"tags": [
					"Project"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "teamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveTeamFromProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProjectDto"
								}
							}
						}
					}
				},
				"tags": [
					"Project"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "teamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/status/list": {
			"get": {
				"operationId": "GetStatuses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStatusListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Status"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/status/{id}": {
			"get": {
				"operationId": "GetStatusById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStatusDto"
								}
							}
						}
					}
				},
				"tags": [
					"Status"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStatusDto"
								}
							}
						}
					}
				},
				"tags": [
					"Status"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IStatusUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Status"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/status": {
			"post": {
				"operationId": "CreateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStatusDto"
								}
							}
						}
					}
				},
				"tags": [
					"Status"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IStatusCreateRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/sprint/list": {
			"get": {
				"operationId": "GetSprints",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISprintListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Sprint"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "boardId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/sprint/{id}": {
			"get": {
				"operationId": "GetSprintById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISprintDto"
								}
							}
						}
					}
				},
				"tags": [
					"Sprint"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateSprint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISprintDto"
								}
							}
						}
					}
				},
				"tags": [
					"Sprint"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISprintUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSprint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Sprint"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/sprint": {
			"post": {
				"operationId": "CreateSprint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISprintDto"
								}
							}
						}
					}
				},
				"tags": [
					"Sprint"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISprintCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/sprint/{id}/start": {
			"post": {
				"operationId": "StartSprint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISprintDto"
								}
							}
						}
					}
				},
				"tags": [
					"Sprint"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/sprint/{id}/complete": {
			"post": {
				"operationId": "CompleteSprint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISprintDto"
								}
							}
						}
					}
				},
				"tags": [
					"Sprint"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/sprint-planning/{planningId}/items/{itemId}/status/{status}": {
			"post": {
				"operationId": "UpdatePlanningItemStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISprintPlanningDto"
								}
							}
						}
					}
				},
				"tags": [
					"SprintPlanning"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/EPlanningItemStatus"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		}
	]
}